integrated plot_generator to coordinator script
